+++ crates/goose-providers/examples/declarative.rs
@@ -13,15 +13,7 @@ async fn main() -> Result<()> {
     let messages = [Message::user().with_text("what is the capital of France?")];
 
     let model = ModelConfig::new("deepseek-v4-flash");
-    let mut stream = provider
-        .stream(
-            &model,
-            "", // session-id
-            system,
-            &messages,
-            &[],
-        )
-        .await?;
+    let mut stream = provider.stream(&model, system, &messages, &[]).await?;
 
     while let Some((Some(msg), _)) = stream.next().await.transpose()? {
         print!("{}", msg.as_concat_text());
