+++ crates/goose/src/providers/mod.rs
@@ -28,6 +28,7 @@ pub mod codex;
 pub mod codex_acp;
 pub mod copilot_acp;
 pub mod cursor_agent;
+pub mod custom_provider_config;
 pub mod databricks;
 pub mod databricks_auth;
 pub mod databricks_v2;
